Eliminate dynamic dispatch from geometry calls
This seems to be a promising win- I'll try it on materials next
